+++ sandbox/tests/physics/fastphysics.ts
@@ -14,11 +14,13 @@ game.debug.collider.showBounds = true;
 ex.Physics.acc.setTo(0, 300);
 //ex.Physics.dynamicTreeVelocityMultiplyer = 1;
 game.currentScene.camera.zoom = 0.5;
+game.toggleDebug();
 var rocketTex = new ex.ImageSource('missile.png');
 var loader = new ex.Loader([rocketTex]);
 
 function spawnRocket(direction) {
-  var rocket = new ex.Actor({x: 300, y: 200, width: 48, height: 16});
+  var rocket = new ex.Actor({x: 300, y: 200, radius: 48, height: 16});
+  rocket.body.canSleep = true;
 
   rocket.on('preupdate', () => {
     if (rocket.isOffScreen) {
@@ -54,9 +56,10 @@ function spawnRocket(direction) {
   game.add(rocket);
 }
 
-var ground = new ex.Actor({x: 0, y: 0, width: 5, height: 5, color: ex.Color.Black.clone()});
+var ground = new ex.Actor({x: 0, y: 400, width: 5, height: 5, color: ex.Color.Black.clone()});
 ground.body.collisionType = ex.CollisionType.Fixed;
-ground.collider.useEdgeCollider(new ex.Vector(0, 400), new ex.Vector(600, 400));
+ground.collider.useEdgeCollider(new ex.Vector(0, 0), new ex.Vector(600, 0));
+// ground.collider.useBoxCollider(600, 30, ex.Vector.Zero);
 game.add(ground);
 
 var rightWall = new ex.Actor({x: 600, y: 200, width: 10, height: 400, color: ex.Color.Black.clone()});
@@ -92,4 +95,7 @@ game.input.keyboard.on('down', (evt: ex.Input.KeyEvent) => {
   }
 });
 
-game.start(loader);
+game.start(loader).then(() => {
+  game.currentScene.camera.pos.x = game.halfDrawWidth;
+  game.currentScene.camera.pos.y = game.halfDrawHeight;
+});