+++ sandbox/tests/physics/physics.ts
@@ -53,11 +53,13 @@ function spawnBlock(x: number, y: number) {
   var block = new ex.Actor({
     pos: ex.vec(x, y),
     color,
-    anchor: ex.Vector.Half
+    anchor: ex.Vector.Half,
+    width: width / 2,
+    height: width + 100
   });
   block.rotation = globalRotation;
   // block.body.addBoxCollider(width + 200, width / 2);
-  block.collider.useBoxCollider(width / 2, width + 100);
+  // block.collider.useBoxCollider(width / 2, width + 100);
   block.body.events.on('contactstart', (e) => {
     // console.log(e);
   });
@@ -78,11 +80,11 @@ function spawnBlock(x: number, y: number) {
 function spawnCircle(x: number, y: number) {
   var width = ex.Util.randomInRange(20, 100);
   var color = new ex.Color(255, ex.Util.randomIntInRange(0, 255), ex.Util.randomIntInRange(0, 255));
-  var circle = new ex.Actor({x: x, y: y, width: width, height: width, color: color});
+  var circle = new ex.Actor({x: x, y: y, radius: width / 2, color: color});
   // circle.rx = ex.Util.randomInRange(-0.5, 0.5);
   circle.angularVelocity = 1;
   circle.vel.setTo(0, 300);
-  circle.collider.useCircleCollider(width / 2);
+  // circle.collider.useCircleCollider(width / 2);
   circle.body.collisionType = ex.CollisionType.Active;
   circle.draw = (ctx: CanvasRenderingContext2D) => {
     ex.Util.DrawUtil.circle(ctx, 0, 0, width / 2, color, color);