+++ src/engine/Collision/Solver/RealisticSolver.ts
@@ -5,12 +5,9 @@ import { CollisionType } from '../CollisionType';
 import { ContactConstraintPoint } from './ContactConstraintPoint';
 import { Side } from '../Side';
 import { Physics } from '../Physics';
-import { Line } from '../../Math/line';
-import { Vector } from '../../Math/vector';
-import { CircleCollider } from '../Shapes/CircleCollider';
-import { ConvexPolygon } from '../Shapes/ConvexPolygon';
 import { CollisionSolver } from './Solver';
 import { BodyComponent } from '../BodyComponent';
+import { CollisionJumpTable } from '../Shapes/CollisionJumpTable';
 
 export class RealisticSolver extends CollisionSolver {
   lastFrameContacts: Map<string, CollisionContact> = new Map();
@@ -191,56 +188,6 @@ export class RealisticSolver extends CollisionSolver {
     }
   }
 
-  private _getSeparation(contact: CollisionContact, point: Vector) {
-    const shapeA = contact.colliderA;
-    const bodyA = contact.colliderA.owner.get(BodyComponent);
-    const shapeB = contact.colliderB;
-    const bodyB = contact.colliderB.owner.get(BodyComponent);
-    if (shapeA instanceof CircleCollider && shapeB instanceof CircleCollider) {
-      const combinedRadius = shapeA.radius + shapeB.radius;
-      const distance = bodyA.transform.pos.distance(bodyB.transform.pos);
-      const separation = combinedRadius - distance;
-      return -separation;
-    }
-
-    if (shapeA instanceof CircleCollider && shapeB instanceof Line) {
-      // TODO circle line separation
-      // return bodyB.getSeparation(bodyA);
-    }
-
-    if (shapeA instanceof Line && shapeB instanceof CircleCollider) {
-      // TODO circle line separation
-      // return bodyA.getSeparation(bodyB);
-    }
-
-    if (shapeA instanceof ConvexPolygon && shapeB instanceof ConvexPolygon) {
-      if (contact.info.localSide) {
-        let side: Line;
-        let worldPoint: Vector;
-        if (contact.info.collider === shapeA) {
-          side = new Line(bodyA.transform.apply(contact.info.localSide.begin), bodyA.transform.apply(contact.info.localSide.end));
-          worldPoint = bodyB.transform.apply(point);
-        } else {
-          side = new Line(bodyB.transform.apply(contact.info.localSide.begin), bodyB.transform.apply(contact.info.localSide.end));
-          worldPoint = bodyA.transform.apply(point);
-        }
-
-        return side.distanceToPoint(worldPoint, true);
-      }
-    }
-
-    if (
-      (shapeA instanceof ConvexPolygon && shapeB instanceof CircleCollider) ||
-      (shapeB instanceof ConvexPolygon && shapeA instanceof CircleCollider)
-    ) {
-      if (contact.info.side) {
-        return contact.info.side.distanceToPoint(bodyA.transform.apply(point), true);
-      }
-    }
-
-    return 0;
-  }
-
   /**
    * Iteratively solve the position overlap constraint
    * @param contacts
@@ -260,7 +207,7 @@ export class RealisticSolver extends CollisionSolver {
           const constraints = this.idToContactConstraint.get(contact.id) ?? [];
           for (const point of constraints) {
             const normal = contact.normal;
-            const separation = this._getSeparation(contact, point.local);
+            const separation = CollisionJumpTable.FindContactSeparation(contact, point.local);
 
             const steeringConstant = Physics.steeringFactor; //0.2;
             const maxCorrection = -5;